feat: optimization to reduce number of database calls, reorganize persistence layer modules
